BUG#18338321 : mysqldbcopy throws python internal errors
The mysqldbcopy utility throws an internal Python exception when used
with --rpl=slave option and  without specifying the --rpl-user. This patch
handles that by fixing variable initialization and it also fixes the help
message for the --rpl option.

This patch also removes the MySQL version restriction on the copy_db_rpl and
copy_db_rpl_errors tests as well as an issue with port handling on mutlib.

Furthermore, it moves the spawn_server method from the rpl_admin test
to the ServerList class from mutlib.

The copy_db_rpl test was modified to exercise this issue.