BUG#18241174: indexcheck Option to report on tables w/o PRIMARY or UNIQUE key
mysqlindexcheck was lacking of an option to report tables without a PRIMARY 
key or UNIQUE key, in the past the absent of indexes or a primary keys were 
only shown when a high verbosity level was used. This patch adds a new option 
to report when a table is not indexed or does not have an UNIQUE index or 
PRIMARY key. In Addition the show indexes option has been improved to show if 
an index is unique and if any of his columns accepts nulls, these columns are 
only shown when verbosity option is also given.

New regression tests were added.
v1