BUG#18483844 : index 'span' for mysqldbcompare
The temporary table created by the mysqldbcompare utility was missing an index
for the 'span' column, meaning that the _COMPARE_SUM and _COMPARE_DIFF queries
were doing full table scans.

This patch fixes this issue by adding the missing index to the temporary
compare table. In addition, a few more changes were made to improve the
performance of the mysqldbcompare utility:
- The BINARY type was used instead of CHAR to store hash values in a more
  optimized way;
- Unnecessary keys were removed from the temporary table (including the
  primary key);
- Use ENGINE=MyISAM for the temporary compare table; 
- A new step was added to the data consistency check, now executing a CHECKSUM
  TABLE first (which is faster when no differences are expected). The current
  comparison algorithm is only execute if this preliminary full table checksum
  fails. A new --skip-checksum-table option was added to skip this new step
  that otherwise is executed by default.

Existing MUT tests were updated accordingly to accommodate the new table
checksum step, namely: the output of the mysqldbcompare utility has changed;
when the --use-indexes option is used the new checksum step must be skipped
otherwise the current algorithm at the row level is skipped if there are no
differences; the minimum width to show the output results correctly aligned is
now 60; the order in which some results were reported has changed.