BUG#17957509 : ReadMe and License files are lost after OSX installer finish
The installation wizard for mac OSX shows the ReadME and License files,
but after finish the installation, these files was not being installed no
where. This patch fixes this issue, by adding duplicates of these files to
the files to be extract and installed in addition of those included used
by the wizard. The extracted files will be located at folder:
   \user\local\mysql-utilities-<version>\ 

No MUT test created.
Tested Manually.
v1