WL#7605 : Add SSL Support
This patch adds the following options to each Utility which uses the --server
or similar option to acquire the connection information.

--ssl-ca=file_name     Path to a list of trusted SSL certificate authorities.
--ssl-cert=file_name   SSL certificate file to use.
--ssl-key=file_name    SSL key file in PEM format.

This patch also adds the ability to read the connection configuration from a
configuration file using the config parser from C/Py.

To use the config-path login mechanism, specify the path of the configuration
file and the group name used to read the connection configuration (from the
--server option). Use the format <config-path>[<[group]>]. For example:
--server=ini.cnf[server_3310].

v6b