BUG#69857 : mysqlrpladmin says "health: Slave is not connected to master."
When the mysqlrpladmin utility is unable to get the a list of aliases
associated with the hostname of the machine, it is possible for a scenario to
arise where the utility warns the user that the "Slave is not connected to the
Master", even though that is not correct. This is not something that can be
fixed from within the utility as it is a DNS configuration issue. However, in
such cases, the utility can print an error message informing users that it is
unable to get the list of aliases. This should be enough to hint the users.
This patch presents that warning message when the utility is used together with
the verbose option.

No MUT tests were added.