BUG#18730720 : mysqluserclone shows grants twice
When using the --dump option on mysqluserclone together with a user whose
hostname is '%', the output from the utility would be repeated, i.e, each grant
would be shown twice. This happens because when checking for grants, the
utility replaces the host part with % to check if the user has any global
privileges, however it does so even if the host is already %.

This patch fixes this issue by checking if the hostname is already %, in 
which case it doesn't try to get the grants again.

A new test was added that exercises this issue.