BUG#18914270 : "ImportError: cannot import name OrderedDict" in utilities 1.5.0
When running some utilities using python v2.6 (the minimum supported
version) an import error was being thrown because of the use of
OrderedDict. The options_parser module was making use of the this
collection, however the OrderedDict was only introduced in Python v2.7.

This patch fixes this issue by removing the OrderedDict dependency
from the module, replacing it with a regular dictionary.
An existing deprecation warning introduced in the Server module was also
removed.