BUG#18790134 : --no-headers do not work with --format=vertical
The --no-headers options does not work with the 'vertical' format. This is not
a bug, it is clearly referred in the documentation that this option only
applies to specific formats (not including the 'vertical' format). 

Nevertheless, this patch improves the description of the --no-headers option
in the help, now explicitly indicating the supported formats for each utility.

Existing MUT tests were updated accordingly.