BUG#18379971: mysqlfailover consumes excessive cpu
The mysqlfailover utility was consuming excessive CPU when executed in console
mode, without using --daemon. The high CPU usage was due to the process to
wait for user input.

This patches fixes this issue by introducing a small idle period (10 ms)
during the polling process to wait for user input, sparing the CPU from an
unnecessary high usage. The fix is applied at a different level for Windows
and POSIX systems, since detection of key hits is handled differently
depending on the operating system.

No new MUT test was added because it requires monitoring the CPU usage
(not supported).