BUG#18790361 : fatal error using mysqlserverinfo as root
When running the mysqlserverinfo utility as root against an offline server,
if fails to start a read-only instance of the server with the fatal error:
'Please read "Security" section of the manual to find out how to run mysqld
as root!'. This error occurs because the mysqld option --user=root is not
used when starting the server. This option is mandatory when starting mysqld
as root.

This patch fixes this issues by detecting the user_name of the user executing
the utility and adding the appropriate --user=user_name option to start the
read-only instance of the server. Additional error checks were added to
validate the --basedir and --datadir options, checking the existance of the
specified directory and access privileges.

New MUT test cases were added to check errors for the --basedir and --datadir
options. However, no new MUT test was added to check the main issue because it
would require running the test as root (not supported).