BUG#18815545 : error: unable to communicate with new instance
This patch implements a change in the way errors are reported when
connections to the server fail. Currently, the code throws a custom
error which includes only a partial representation of the actual
error returned.

Rather than send a custom error, the code now sends (raises) whatever
error is returned during the connection. Not only does this correct
a deviant behavior, it makes the error more standard with the other
MySQL client applications.

Note that this patch does not address the specific error as described
in the bug report. This is because the problem reported isn't a bug.
The user reported strange errors when trying to connect to downed
servers. The error presented before this patch is mostly correct,
but I felt the code could be improved by reporting the true and
accurate error as received from the connection.

Thus, this patch is related to the bug report but only because it
makes the errors reported more clear.

CAB v2