BUG#18730720 : mysqluserclone shows grants twice
This patch fixes a test regression issue added by the previous patch
for this bug.

The clone_user test had a slightly different output on Windows
systems, using single quotes around the hostname which are not present
on Unix systems.

This patch adds a mask to the test case in order to fix that issue.