BUG#18768780 : mysqluserclone creates user on destination host, after error
When cloning a user, the mysqluserclone utility was not checking if
the user of the destination server had enough privileges for a
successful operation.

This patch fixes this issue, by checking if the user of the
destination server has a superset of the privileges of the user being
cloned.

This patch also improves the User class by adding some new instance variables
to store (cache) the grant privileges. Now instead of reading the grant 
privileges from the server each time, we can consult the cache, saving some
time and reducing server communication. 
This patch also fixes existing pylint issues.

Existing MUT tests were modified to exercise this issue and a new unit test
was added.