BUG#18852960 : mysqluc types qxit when type q + enter
This patch corrects a defect in handling aliases. Currently, the code
recognizes an alias for any command. Thus, when typing the first few
letters and a tab, it will complete the command. However, it failed to
recognize and detect aliases.

Since 'quit' is an alias for 'exit', when a 'q' followed by a tab is
entered, the code returned 'qxit', which would fail if you pressed
ENTER. This patch allows the code to detect alias so that when you repeat
this scenario, the code returns 'quit'.

Unit test updated for regression testing.

CAB v2