BUG#19009690 : AttributeError using mysqlfailover utility
When a login-path (or config-path) is used for the server connection
values and the file '.mylogin.cnf' is not found, an AttributeError was issued
because the error string for the config-path was incorrectly handled. This
patch fixes this issue removing the invalid error message attribute. In
addition, a letter case issue was fixed (first letter converted to lowercase)
for the config-path error message in order to use the correct case in the full
error message.

A new unit test was added for regression. Additionally, a few MUT tests were
updated for the letter case issue.