BUG#18768780 : mysqluserclone creates user on destination host, after error
This patch fixes a test regression issue added by the previous patch
for this bug. The previous patch replaced the 'GRANT OPTION' privilege
with 'WITH GRANT OPTION'. This patch fixes that issue, making use of
'GRANT OPTION' again as the name of the privilege.

An existing test output was updated to reflect this change.