BUG#18730815 : mysqluc suddenly exits when a command misses closing quotation
The mysqluc utility suddenly exits raising an error when a command is 
wrongly formatted, i.e. when it misses the closing quotation. This patch 
fixes the error by providing a warning to the user instead of exiting from 
the console, letting the user edit the command.

A MUT tests case was added.

v2