BUG#18815523 : unclear connection message --discover-slaves-login.
When using the --discover-slaves-login option (for example with
mysqlrplsync) and there are connection or configuration issues with
some of the slaves, those slaves are skipped and only a generic
warning message is issued. The same behavior is observed even with the
--verbose option instead of indicating specifically which slaves were
skipped and why.

This patch fixes that issue by providing clearer information to
notify the user when any connection or configuration issue is
detected during the discovery process. A simple warning is issued by
default, clearly indicating that the user needs to use the --verbose
option for more details. When the --verbose option is used, specific
information for each slave and respective error is displayed.

Existing MUT tests were updated and new tests added (using --verbose)
for regression. In addition, some masks were also added to a few
existing tests in order to fix non deterministic error messages
(platform dependent).