BUG#19178471 : Error using mysqlreplicate or mysqlrplms without --rpl-user
The mysqlreplicate and mysqlrplms utilities were raising a Python exception
when the --rpl-user option was not provided. This is a regression issue that
was introduced by a previous patch.

The --rpl-user option is now mandatory for these utilities.

MUT tests were added for regression and existing ones were corrected
according to this change.