BUG#19076488 : mysqlrpladmin does not check errant transactions
The check for errant transaction was not performed for the switchover command
of the mysqlrpladmin utility, only for the failover command. This patch adds
the errant transaction check also to the switchover command for the
mysqlrpladmin utility.

Note: The errant transaction check requires all servers in the topology to
have GTID enabled, otherwise a warning is issued indicating that the check
cannot be performed for the switchover command.

New MUT test cases were added for regression and existing rpl_admin tests
were updated to handle the issued warning when the check cannot be performed.