QA Fix
This patch fixes an issue with our Hudson scripts which were not
installing connector python in a correct way. They were just
copying the connector python folder to the utilities folder, but
recent changes to the utilities code make that approach not
viable.  With these changes we are able to install the connector
python using the setup.py script, which is the correct way to do
it.