BUG#19262669 : fixes to Package mechanisms regarding release-1.5.1rc
The release-1.5.1rc will be a standalone release, meaning it Utilities will
no longer be a subproduct of Workbench. As a result we need to fix some of
the Package mechanism.

This patch fixes the Package mechanism to remove the references to Workbench,
in addition adds the functionality to package additional info files from
Fabric proyect.

This patch also fixes the shortcuts to the online documentation installed by
the Windows MSI package.

v3