BUG#19238432 : utilities do not parse proxy and stored routines grant statements
The current utilities code did not parse all types of grant statements.
For instance, it didn't work with PROXY privileges, stored routines
and it also had an issue with objects with backticks.

This patch fixes all the above issues by improving the existing regular
expression.

An existing unit test was modified to exercise these issues.