BUG#14163181 : mysqlprocgrep needs to show killed processes
Currently, the mysqlprocgrep --kill option silently kills processes
and connections without reporting which are killed. The original
workflow and use of the utility is such that the user runs first with
the --sql option, uses the output to run on a live server, then issues
the kill.

However, this feature request is to add a display of the processes
and connections killed during the --kill operation. The utility will
now display those rows from the SHOW PROCESSLIST that match the query
and are killed.

v2