BUG#18543913 :  allow health report for list of slaves
Currently, the health report requires the master connection. If the user
wants to see a health report for a list of slaves, the utility throws an
error because the master is not specified.

This feature request adds the ability to see the health report for a
list of slaves with or without the master specified.  If no master is
specified, the health column for the slave will state "No master
specified.". This new option works with verbose mode as well.

v2