BUG#18632836 : --skip-table-options does not work when difftype=sql
When table options are skipped in mysqldiff and the difference type
is SQL, the utility failed to skip said table options. This is only
true when the tables compared have unequal table options. For example,
the first table has an option the second does not.

This patch corrects the problem by checking for the skip during the
transformation generation. A test case was added for regression.

CAB v2