BUG#19320164 : mysqldbcompare --disable-binary-logging not working properly
The disable-binary-logging flag in mysqldbcompare was not working
properly, even when not used it was still enabled internally. Also,
the utility was throwing an error when trying to generate SQL
statements for the differences between tables if one of the rows had
columns with NULL values. Furthermore, the documentation needs to be
improved in order to correctly show the required privileges necessary
to run the utility as well as to clearly indicate that the utility
is intended for "warm" usage,i.e, databases being compared must not be
modified during comparison.
This patch fixes all of the above issues and also adds calls the ability
to call the to_sql method for each column, which is enabled by default
for csv output.

This bug fixes BUG#19174364 as well as BUG#71789.

A new test case was introduced and existing ones were updated.