WL#6187 : Binary log relocate
This patch adds the new mysqlbinlogmove utility to move existing binary log
files (binlog or relay logs) to a new location, updating the corresponding
index file accordingly. In particular, it allows two main operations; 1)
archive older binary logs to a new location with the server running (not
moving the latest binary log) using the --server option, and 2) relocate all
binary logs with the server stopped to a new location using the --binlog-dir
option (for  example, allowing the server option --log-bin to be changed
afterwards without errors when starting the server).

In addition, some improvements were made to the existing code, namely:
- Corrected the error message when an invalid end date format is used for the
  mysqlauditgrep utility;
- Improve check_dir_option() in the mysql.utilities.common.option module to
  return the absolute path of the specified directory.
- Add functions get_value_intervals_list() and check_date_time() to the
  mysql.utilities.common.option module to improve code reuse when checking
  option values with interval lists and date/time.
- Add a new message PARSE_ERR_OPTS_EXCLD template to improve error message
  reuse for option that cannot be used simultaneously;

New MUT tests were added for the new utility, as well as a new unit test for
the new common module.