WL#6187 : Binary log relocate
The binlogmove_parameters test fails on Windows due to a server issue for
the SHOW BINARY LOGS statement. This statement display unexpected and wrong
information after relocating the binary log files if it is not executed
previously.

This patches adds a workaround for this regression issue, executing a first
SHOW BINARY LOGS on Windows before relocating the binary log files. This way
the binlogmove_parameters test runs correctly on all platforms.

In addition, a new method was added to the binlogmove_privileges test in
order to ensure that all relay log files are created before running the test
case for the slave, otherwise it might fail in some situations.