BUG#19675964 : README file name in DMG packages is not congruent to other pkgs
The README file in the MySQL Utilities commercial DMG packages is being named
ReadMe.txt instead of README_com.txt unlike in other packages. Similarly, the
LICENSE file was packed as License.txt instead of LICENSE.txt. This patch
fixes the packaging code to correctly include these files with names that are
congruent with the other packages.

The license notice that appears in the Python egg for the commercial 
distribution has also been corrected to be NO GPL applicable.

As part of fix BUG#19675930, the bdist commands has been updated to include
the README_Fabric_com.txt instaed of README_Fabric.txt

v3 patch for mu-1.5.2