BUG#19430737: missing mysqldbcompare feature to compare all databases
The documentation for the mysqldbcompare utility refers some options (i.e.,
--all, --exclude and --regexp) related with the comparison of all databases
between two servers that are not available. These options and part of the
related code were accidentally removed in a previous merge between releases.

This patch recovers the missing code and features (from BUG#14059317) to
allow mysqldbcompare to check all databases between two servers. A few
improvements were also added, namely error checking for the required use of
the --server1 and --server2 options, and verify if the different servers are
used to compare all databases.

MUT tests were added for regression.