BUG#18718660 : indexes that just have same prefix are not redundant
In MySQL a BTREE index A is considered redundant of a BTREE index B
if and only if A is a leftmost prefix of B. Having a common partial
prefix is not enough. The mysqlindexcheck utility had an issue and was
considering to BTREE indexes to be redundant if they had a common
partial prefix. This patch fixes that issue and the utility now correctly
checks if a BTree index is a leftmost prefix of other.

A new test case was added to exercise this issue.