BUG#19313139 : mysqldbcompare not escaping values in sql output
When comparing databases, quote characters within resulting SQL
transformation statements are not escaped. As a result, the
statements are not valid and cannot be executed directly in a
client tool, as intended.

This patch fixes this issue by correctly escaping quote characters.
A new test case was added to correctly exercise this issue.

The warning message that is printed when the utility does not present
changes was also improved by this patch.