BUG#18763262 : error using large span key values for mysqldbcompare
The mysqldbcompare utility fails to execute and issues a query error when
when a large value is used for the --span-key-size option. This error occurs
because the utility tries to store a value too big in the temporary compare
table for the span key column.

This patch solves this issue by adding an error check for the --span-key-size
option to ensure that the specified value is smaller or equal than the
maximum allowed value. The maximum allowed value for the span key size is 32
since it is a substring of the hash key with size 32.

In addition, a warning is issued if odd values are used for the span key size.
Internally, the hash data in the temporary compare table is converted
(compressed) to binary for performance reasons, stored in a field with half
it original size. Therefore, the effectively used span key size must be an
integer multiple of 2 (even number).

MUT tests were added for regression.