BUG#19781888 : MUT test binlogmove_privileges failing
The MUT test binlogmove_privileges test fails on some servers due to the
unexpected rotation of relay log files on the slave. One of the reasons for
a slave to decide to create a new relay log might be the slave IO thread
reconnecting.

This patches fixes the affected binlogmove MUT tests by stopping all slaves
as soon as the test relay log files are created. This way no further relay
log rotation is performed due to the reconnection of the slave IO thread to
the master.