BUG#19720715 : mysqlindexcheck has wrong results for hash and fulltext indexes
The results regarding index redundancy shown by mysqlindexcheck are
not correct for FULLTEXT nor HASH type indexes.

For FULLTEXT indexes the order is not relevant, nor is the number of
columns.  As for HASH type indexes, one index in considered to be
redundant of another index,if and only if it contains the exact same
columns with the exact same order, i.e, two HASH type indexes are
redundant only if they are duplicate.

This patch fixes the behavior of mysqlindexcheck, which now uses
the rules described above for identifying redundant HASH and FULLTEXT
indexes.

Existing tests were modified in order to exercise this issue and prevent
regression issues.