BUG#18717843 : duplicated execution for mysqldbcompare with --show-reverse
The mysqldbcompare utility was executing twice the full algorithm to search
for data differences when the --show-reverse option was used, instead of only
repeating the part that generate the differences in the reverse direction.
This can have a notable impact in terms of performance for larger databases.

This patch fixes this issue by changing the code structure. The logic of the
reverse feature was moved from the function _check_data_consistency() in
the command.dbcompare module to the function check_consistency() in the
common.dbcompare module, in order to avoiding the unnecessary repetition
of the full algorithm to find the differences between the tables.

No new MUT test was added. Nevertheless, some test result files were updated
to remove incorrect duplicated outputs previously resulting from this issue.