BUG#19781888 : MUT test binlogmove_privileges failing
Sometimes the MUT test binlogmove_privileges test crashes during the cleanup
process on Windows, failing to delete the temporary directory to move files.
This issue occurs in a non-deterministic way and might be due to one of the
test server still using a moved binary log file while the directory is being
removed.

This patch fixes this issue by switching the operation to kill the test
servers with the one to remove the temporary directory. Killing the servers
first will ensure that no file will still be in use while removing the
temporary directory.