BUG#19238195 : Missing support of config-path in rpl_user option.
The rpl_user option does not support the use of config-path for the login
values. This patch changes the current behavior used to parse the login
values given for the replication user given with the rpl user option by
adding the missing support for config-path that was not previously added
to this option.

A regression test case was added to exercise the change.
v2