BUG#19030588 : ssl-ca option is required to establish an encrypted connection
All the utilities were requiring the '--ssl-ca' option in order to establish 
an encrypted connection to the server, but this behavior differs of the mysql 
client where this option is optional if the '--ssl-cert' and '--ssl-key' 
options are specified. This was being occurring because connector python does 
not accepts a 'None' value for ssl-ca option that was being used in the 
connection parameters in the case no other value was specified by the user 
while other ssl option was specified. This patch changes this wrong behavior 
by setting an empty string instead of the 'None' value when the '--ssl-cert' 
and '--ssl-key' options are specified.