BUG#19792492 : Commercial mysql utilities rpm depends on community c/py.
The commercial rpm distribution package for MySQL Utilities is requiring the 
rpm community version of connector python as an install prerequisite instead 
of the commercial version. This is a regression that was introduced at some 
point of time when the rpm package descriptor was changed to produce a second 
package for the Fabric Doctrine. This patch fixes this issue by correcting 
the install prerequisite to require the commercial version of C/py.

Note, The RPM can not hold an install prerequisite of type 'OR', where a 
required package is A OR B as the Debian distribution package does, this is 
the main reason for the Utilities rpm package to require the connector python 
rpm package with the same license type to be installed.

v1