BUG#19942512 : utilities do not work with spaces on the mysql server path
MySQL Utilities do not correctly handle spaces in the path to the
MySQL server. This is a common case on Windows. As a result utilities
such as mysqlserverinfo or mysqlserverclone do not work properly under
such circumstances.

This patch fixes this issue by surrounding the paths used by these utilities
with quotes.

The clone_server and clone_server_parameters  tests were modified to test
this issue.