WL#7350 : mysqlslavetrx
This patch adds the new 'mysqlslavetrx' utility to skip transaction on
slaves. More precisely, empty transactions are inject on the specified slaves
for the given GTID set.

In addition, some improvements were made to the existing code, namely:
- A spelling issue (missing space) was fixed in the description of the
  --slaves option for the help.
- A new parameter was added to the check_privileges() function in the
  mysql.utilities.common.topology module in order to check the salves and
  skip the master.
- A new common module 'gtid' was created to hold all function to manipulate
  GTIDs. Existing functions from the replication module (get_last_server_gtid,
  gtid_set_cardinality, gtid_set_union) were moved to the new module and a
  new one was added (gtid_set_itemize).

New MUT tests were added for the new utility, as well as a new unit test for
the new common module.