BUG#19781304 : possible to execute system commands with mysqluc
The mysqluc utility allows users to execute (inject) system commands by
appending a valid command separator character such as ';', '||' or '&&'
to a valid mysqluc command.
This vulnerability is called shell injection and is present in mysqluc as well
as mysqlfailover and mysqlrpladmin. To prevent it, this patch replaced all
shell=True ocurrences to shell=False made in subprocess.Popen constructors
where input could be directly controlled by the user.

Furthermore in both mysqlfailover and mysqlrpladmin, an error message is now
printed in case any of the external script files provided via the existing
options does not exist or is not executable. Also, due to changing
the shell value to False, on unix it is required for the external scripts to
have a header (shebang) indicating the path to the interpreter that will run
the script.

This patch also improves the kill_server method from the mutlib module to
make it more resilient to possible exceptions.

Some tests were added in order to correctly exercise these issues,  preventing
regression changes.