BUG#20035380 : utilities need to support new mysql fk behavior
Starting from MySQL 5.5.41 and 5.6.22, there is a new behavior with foreign
keys.
Now you need to explicitly indicate the schema to which the table that the
foreign key points to belongs. Previously, if missing, it was assumed that
the schema of the table pointed by the foreign key was the schema of the table
being defined, but starting with 5.5.41 and 5.6.22 that is no longer true.

This patch fixes this issue by using full qualified names
(i.e., database.table) for table statements that are not directly retrieved
from the server.
In addition, in the std_data/fkeys.sql file it fixes minor data inconsistencies
and it adds USE statements in order to correctly load the file for MySQL
servers with the new behavior.

No new MUT test was added because the import_basic and import_parameters tests
already cover this situation (previously failing with the new behavior).