WL#8039 : MySQL binary log purge and rotate utilities
This patch adds the new mysqlbinlogpurge and mysqlbinlogrotate utilities
to purge and rotate binary logs.

The mysqlbinlogpurge operates in a safe manner prohibiting deletion of binary
log files that are open or those required by slaves (have not been read by
the slave).

The mysqlbinlogrotate rotates the binary log file that the server is
currently writing to by closing it and creating a new binary log file. The
rotation can be conditional on the size of the active binary log by using the
--min-size option.