BUG#19031182 : Utilities does not verify ca certificate.
The utilities behavior when verifying the SSL ca certificate was not consistent
with the one from MySQL client when the the '--ssl=1' option is provided. This
patch adds a new '--ssl' option to indicate that the ca certificate must be
verified. In addition the '--ssl' option can be used to establish an encrypted
connection to a server configured to use SSL without need to specify any SSL
certificate.

Note: With this patch the behavior mentioned on WL#6791 is shared by Utilities
and MySQL client regarding ca cert validation: there will
be no guarantee for server certificate validation unless --ssl-ca also is
specified when using --ssl=1 option.

New test cases were added to exercise the changes.