QA : Release 1.6.1 Testing
This patch corrects small issues with some of MUT's tests.
It updates the content of some result files in order
to support the new --ssl option. It also fixes the list
of privileges of the user created in copy_db and copy_db_errors
test. Furthermore, it changes the startup command of the servers used
in ssl_connection_simple_test_errors in order to obtain a consistent
behavior between commercial and GA versions of the server regarding SSL.

It also removes an unnecessary check from one of the functions from the
Hudson support files.