BUG#20583756 : mysqluc fails to launch utilities when using Windows MSI.
mysqluc was failing to launch the utilities when they had been installed
using the Windows MSI. This error was due to a Windows incompatibility 
with the use of both absolute paths and shell=False. With this fix, the 
subprocesses are not longer created using the full path of the 
executables, instead only the name is used for the previously described 
scenario.