BUG#20088285 : Help text for SSL Options should start with lower case.
The help text for the SSL options it was capitalized and it does not have
the same format than the other options. This patch corrects the help text
of all the SSL options that are used in all the utilities and updates the
test results where is necessary. In addition the the option --run-tests-like
has been added to MUT to enable the execution of all tests that contains the
given string with the option.