BUG#20161135 : mysqlrpl_admin_gtid test fails with mysql 5.6.22 and 5.7.6
One of the tests in rpl_admin_gtid was setting the gtid_purged
variable on the master server in order to create a controllable
scenario where a slave server is missing some GTIDs from the
server. However this was a faulty behavior from the server and has
been fixed 5.6.22 and 5.7.6 on BUG#19012085.

With this patch, that test now makes use of empty transactions to
test the same scenario.

The respective result file was updated because the output changed
slightly.