BUG#20344064 : error retrieving slave_master_info data for 5.7.6 servers
The mysqlfailover and mysqlrpladmin utilities were crashing when performing
some operations for MySQL server of version 5.7.6. This issue occurred
because a new column 'Channel_name' was added to the mysql.slave_master_info
table starting from the referred server version and this change was not
supported by MySQL utilities.

This patches fixes the issue by adding the new column 'Channel_name' to the
list of supported columns and improving the function to read the
slave_master_info data in order to not crash if new unexpected columns are
found.

Some additional issues were also fixed for 5.7.6 servers, namely:
- Add missing --no-defaults option for serverclone to start servers with
  version greater than 5.7.6.
- Change queries for the mysql.user table to use the column
  'authentication_string' instead of 'password' (removed starting from 5.7.6).

MUT tests were update for regression, namely a mask for the 'Channel_Name'
data was added and the masks for the rpl_admin test were improved.