BUG#20302446 : mysqlgrants doesnot output users with specific permissions
The mysqlgrants shows all the users that have permission to an object
including also those users that inherit the permission from an upper
level but does not allow specify the level of inheritance. This patch
adds a new option '--inherit-level' to allow the user to reduce the
resulting list of users with the same privileges at object, database or
global levels.

In addition the wildcard "*" can be used along with the database name in
the form "<db_name>.*" and allow mysqlgrants to show the grants for all
the objects in that database instead of list each individual object when
show all objects is desired.

New tests cases were added to exercise the new option and the use of the
wildcard.

v2.