BUG#20756920 : fix failing mut tests on mysql 5.7.6
MySQL server version 5.7.6 introduced some changes that affected MySQL
Utilities. This patch adds new masks to some of the existing tests in
order to make the result files compatible with 5.7.6.

It also fixes an issue with the serverinfo utility, it was using
relative paths in order to obtain the size of the logs, which
does not work.