BUG#20409105 : MUT tests fail with NO_AUTO_CREATE_USER sql_mode
Various MUT tests are failing when the sql_mode is set to
NO_AUTO_CREATE_USER at the server used to run MUT tests.
This is because in some tests user creation is done implicitly
via GRANT statements. We fix this by explicitly calling CREATE
USER before every GRANT statement.

Some tests are still failing, following is the list of failing tests
and the corresponding bug in which they will be fixed:

* copy_db, copy_db_errors, copy_db_locks, copy_db_parameters,
  copy_db_skip_objects copy_db_gtid copy_db_rpl - BUG#20972389
* copy_user - BUG#21031993
* rpl_sync rpl_sync_filters, rpl_sync_parameters, rpl_sync_privileges,
  rpl_sync_ssl, compare_db -  BUG#20979041
* check_rpl_table, rpl_admin_gtid_demote_master_repo_file,
  export_gtid - BUG#21110885

MUT tests are modified with this patch.