BUG#20972389 : Fix mysqldbcopy to copy dataabase to a server with NO_AUTO_CREATE_USER
While copying a database to a server with sql mode NO_AUTO_CREATE_USER
mysqldbcopy was throwing error. The error was occuring due to executing
GRANT statements on the destination server without creating the user.
We fix this by first creating the user if it does not exist before
executing grant statements.

This patch also fixes some of the failing MUT tests when run with
MySQL server 5.7.7

No test case was added as the patch fixes some of the already broken tests.